refactor(client): move customers slice into features/customers
