+++ client/src/app/App.tsx
@@ -17,7 +17,7 @@ const BarcodeTools = lazy(() => import('../pages/BarcodeTools'));
 const Deliveries = lazy(() => import('../pages/Deliveries'));
 const SalesHistory = lazy(() => import('../pages/SalesHistory'));
 const UsersPage = lazy(() => import('../features/admin/pages/Users'));
-const CustomersPage = lazy(() => import('../pages/Customers'));
+const CustomersPage = lazy(() => import('../features/customers/pages/Customers'));
 const DistributorsPage = lazy(() => import('../pages/Distributors'));
 const CategoriesPage = lazy(() => import('../pages/Categories'));
 const SettingsPage = lazy(() => import('../features/admin/pages/Settings'));
@@ -31,10 +31,10 @@ const CustomerDisplay = lazy(() => import('../pages/CustomerDisplay'));
 const RegisterPage = lazy(() => import('../pages/Register'));
 const ShiftsPage = lazy(() => import('../pages/Shifts'));
 const ExpensesPage = lazy(() => import('../pages/Expenses'));
-const SegmentsPage = lazy(() => import('../pages/Segments'));
+const SegmentsPage = lazy(() => import('../features/customers/pages/Segments'));
 const LayawayPage = lazy(() => import('../pages/Layaway'));
 const BundlesPage = lazy(() => import('../pages/Bundles'));
-const FeedbackPage = lazy(() => import('../pages/Feedback'));
+const FeedbackPage = lazy(() => import('../features/customers/pages/Feedback'));
 const BackupPage = lazy(() => import('../features/admin/pages/Backup'));
 const BranchesPage = lazy(() => import('../features/admin/pages/Branches'));
 const StorefrontPage = lazy(() => import('../pages/Storefront'));
