+++ client/src/features/customers/pages/Customers.tsx
@@ -3,10 +3,10 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { Plus, Pencil, Trash2, History, Star } from 'lucide-react';
-import { Button } from '../shared/ui/button';
-import { Input } from '../shared/ui/input';
-import { Textarea } from '../shared/ui/textarea';
-import { Label } from '../shared/ui/label';
+import { Button } from '../../../shared/ui/button';
+import { Input } from '../../../shared/ui/input';
+import { Textarea } from '../../../shared/ui/textarea';
+import { Label } from '../../../shared/ui/label';
 import {
   Dialog,
   DialogContent,
@@ -14,7 +14,7 @@ import {
   DialogTitle,
   DialogFooter,
   DialogDescription,
-} from '../shared/ui/dialog';
+} from '../../../shared/ui/dialog';
 import {
   AlertDialog,
   AlertDialogAction,
@@ -24,11 +24,11 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '../shared/ui/alert-dialog';
-import DataTable from '../shared/components/DataTable';
+} from '../../../shared/ui/alert-dialog';
+import DataTable from '../../../shared/components/DataTable';
 import CustomerDetail from '../components/CustomerDetail';
-import { useTranslation, t as tStandalone } from '../shared/i18n/index';
-import { resource } from '../shared/lib/resource';
+import { useTranslation, t as tStandalone } from '../../../shared/i18n/index';
+import { resource } from '../../../shared/lib/resource';
 import type { ColumnDef } from '@tanstack/react-table';
 import type { Customer } from '@/types';
